pyproject.toml
[tool.poetry]
name = "nua-infra-abilian"
version = "0.1.0"
description = ""
authors = ["Stefane Fermigier <sf@fermigier.com>"]
#readme = "README.md"
#packages = [{include = "nua_infra_abilian"}]

[tool.poetry.dependencies]
python = "^3.9"

invoke = "^2"

# Not needed currently, as nua-build is only useful on the same machine
# as the nua-orchestrator.
# nua-build = { git = "https://github.com/abilian/nua.git", branch = "main", subdirectory = "nua-build" }
toml = "^0.10.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
devtools = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
